+++ TablePro/Views/Results/Cells/DataGridCellPalette.swift
@@ -10,6 +10,11 @@ struct DataGridCellPalette: Equatable {
     let regularFont: NSFont
     let italicFont: NSFont
     let mediumFont: NSFont
+    let text: NSColor
+    let placeholderText: NSColor
+    let booleanTrueText: NSColor?
+    let booleanFalseText: NSColor?
+    let rowNumberText: NSColor
     let deletedRowText: NSColor
     let modifiedColumnTint: NSColor
     let findMatchTint: NSColor
@@ -18,6 +23,11 @@ struct DataGridCellPalette: Equatable {
         regularFont: .systemFont(ofSize: NSFont.systemFontSize),
         italicFont: .systemFont(ofSize: NSFont.systemFontSize),
         mediumFont: .systemFont(ofSize: NSFont.systemFontSize, weight: .medium),
+        text: .labelColor,
+        placeholderText: .secondaryLabelColor,
+        booleanTrueText: nil,
+        booleanFalseText: nil,
+        rowNumberText: .secondaryLabelColor,
         deletedRowText: .secondaryLabelColor,
         modifiedColumnTint: .systemYellow,
         findMatchTint: .findHighlightColor
@@ -30,6 +40,11 @@ extension ThemeEngine {
             regularFont: dataGridFonts.regular,
             italicFont: dataGridFonts.italic,
             mediumFont: dataGridFonts.medium,
+            text: colors.dataGrid.text,
+            placeholderText: colors.dataGrid.nullValue,
+            booleanTrueText: colors.dataGrid.boolTrue,
+            booleanFalseText: colors.dataGrid.boolFalse,
+            rowNumberText: colors.dataGrid.rowNumber,
             deletedRowText: colors.dataGrid.deletedText,
             modifiedColumnTint: colors.dataGrid.modified,
             findMatchTint: .findHighlightColor